+++ Display/Controls/UserInfo.xaml
@@ -14,21 +14,24 @@
         <StackPanel
             Orientation="Horizontal"
             Padding="20,0"
-            Visibility="{x:Bind isShowUserInfo(status),Mode=OneWay}">
+            Visibility="{x:Bind IsShowUserInfo(Status),Mode=OneWay}">
             <!--头像-->
-            <PersonPicture ProfilePicture="{x:Bind userinfo.face,Mode=OneWay}" x:Name="face_Picture"/>
+            <PersonPicture ProfilePicture="{x:Bind Userinfo.face,Mode=OneWay}" x:Name="FacePicture"/>
 
             <!--信息-->
-            <StackPanel x:Name="userInfo_StackPanel" Padding="5,5" HorizontalAlignment="Center">
-                <StackPanel x:Name="vipCard_stackpanel"
+            <StackPanel x:Name="UserInfoStackPanel" Padding="5,5" HorizontalAlignment="Center">
+                <StackPanel x:Name="VipCardStackPanel"
                             Background="#ffa502"
                             CornerRadius="5"
                             Width="50"
                             HorizontalAlignment="Right"
-                            Visibility="{x:Bind isShowVip(status),Mode=OneWay}">
-                    <TextBlock FontSize="10" FontFamily="Segoe UI" Foreground="White" HorizontalAlignment="Center" Text="{x:Bind data:UserData.getVipName(userinfo.mark1),Mode=OneWay}"/>
+                            Visibility="{x:Bind IsShowVip(Status),Mode=OneWay}">
+                    <TextBlock FontSize="10" FontFamily="Segoe UI"
+                               Foreground="White"
+                               HorizontalAlignment="Center"
+                               Text="{x:Bind data:UserData.getVipName(Userinfo.mark1),Mode=OneWay}"/>
                 </StackPanel>
-                <TextBlock x:Name="UserName_TextBlock" FontWeight="Bold" FontSize="20" HorizontalAlignment="Center" Text="{x:Bind userinfo.user_name,Mode=OneWay}"/>
+                <TextBlock x:Name="UserNameTextBlock" FontWeight="Bold" FontSize="20" HorizontalAlignment="Center" Text="{x:Bind Userinfo.user_name,Mode=OneWay}"/>
                 <!--<TextBlock FontWeight="Bold" FontSize="20" HorizontalAlignment="Center""/>-->
                 <HyperlinkButton HorizontalAlignment="Right" Content="更新信息" Click="UpdateInfoButton_Click"/>
                 <HyperlinkButton HorizontalAlignment="Right" Content="退出登录" Click="Logout_Click"/>
@@ -37,16 +40,16 @@
         </StackPanel>
 
         <!--无数据时-->
-        <Grid Visibility="{x:Bind isShowOtherInfo(status),Mode=OneWay}" >
+        <Grid Visibility="{x:Bind IsShowOtherInfo(Status),Mode=OneWay}" >
             <StackPanel VerticalAlignment="Center">
                 <!--未登录时-->
-                <StackPanel Visibility="{x:Bind isProcess(status),Mode=OneWay}" Orientation="Horizontal" Spacing="5">
+                <StackPanel Visibility="{x:Bind IsProcess(Status),Mode=OneWay}" Orientation="Horizontal" Spacing="5">
                     <TextBlock Text="未登录，请先登录" VerticalAlignment="Center"/>
                     <Button Content="登录" Click="LoginButton_Click"/>
                 </StackPanel>
                 
                 <!--更新信息时-->
-                <StackPanel Orientation="Horizontal" Spacing="10"  Visibility="{x:Bind isUpdate(status),Mode=OneWay}">
+                <StackPanel Orientation="Horizontal" Spacing="10"  Visibility="{x:Bind IsUpdate(Status),Mode=OneWay}">
                     <ProgressRing IsActive="True"/>
                     <TextBlock Text="正在更新信息" VerticalAlignment="Center"/>
                 </StackPanel>
