+++ Display/Models/Settings/NavLink.cs
@@ -1,0 +1,11 @@
+﻿using System;
+
+namespace Display.Models.Settings;
+
+internal record NavLink
+{
+    public string Label { get; init; }
+    public string Glyph { get; init; }
+
+    public Type NavPageType { get; init; }
+}
