+++ Display/Views/Settings/AccountPage.xaml
@@ -1,0 +1,109 @@
+<?xml version="1.0" encoding="utf-8"?>
+<Page
+    x:Class="Display.Views.Settings.AccountPage"
+    xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+    xmlns:local="using:Display.Views.Settings"
+    xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+    xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+    xmlns:interactivity="using:Microsoft.Xaml.Interactivity"
+    xmlns:behaviors="using:CommunityToolkit.WinUI.Behaviors"
+    xmlns:customControls="using:Display.Controls"
+    xmlns:controls="using:CommunityToolkit.WinUI.Controls"
+    xmlns:data="using:Display.Models.Data"
+    mc:Ignorable="d"
+    Background="{ThemeResource ApplicationPageBackgroundThemeBrush}">
+
+    <Page.Resources>
+        <Style TargetType="controls:SettingsCard"
+               BasedOn="{StaticResource BaseSettingsCardStyle}"/>
+    </Page.Resources>
+
+    <Grid RowDefinitions="auto, auto, *">
+
+        <!--用户信息-->
+        <Grid MaxWidth="{StaticResource SettingsMaxWidth}">
+            <customControls:UserInfo x:Name="UserInfoControl"
+                                     UpdateInfoClick="UpdateInfoButton_Click"
+                                     LoginClick="LoginButton_Click"
+                                     LogoutClick="LogoutButton_Click"
+                                     HorizontalAlignment="Left"
+                                     Margin="5" />
+        </Grid>
+
+        <!--Cookie-->
+        <!--<controls:SettingsExpander Grid.Row="1"
+                                    Header="115"
+                                    MaxWidth="{StaticResource SettingsMaxWidth}">
+            <controls:SettingsExpander.Items>
+
+            </controls:SettingsExpander.Items>
+        </controls:SettingsExpander>-->
+
+        <StackPanel Grid.Row="1"
+                    Style="{StaticResource SettingStackPanel}">
+            <controls:SettingsCard Header="Cookie"
+                                   HorizontalContentAlignment="Stretch"
+                                   ContentAlignment="Vertical">
+                <RelativePanel>
+                    <PasswordBox x:Name="CookieBox"
+                                 MinWidth="100"
+                                 MaxWidth="600"
+                                 Password="{x:Bind data:AppSettings._115_Cookie}" />
+
+                    <StackPanel Orientation="Horizontal"
+                                RelativePanel.AlignRightWithPanel="True"
+                                Spacing="5">
+                        <Button Content="删除"
+                                Click="DeleteCookieButton" />
+                        <Button Content="复制"
+                                Click="CopyCookieButtonClick" />
+                        <Button Content="导出"
+                                Click="ExportCookieButton">
+                            <ToolTipService.ToolTip>
+                                <TextBlock>
+                                     配合浏览器插件(Cookie Edit)使用
+                                </TextBlock>
+                            </ToolTipService.ToolTip>
+                        </Button>
+                        <ToggleButton Content="显示"
+                                      Click="Show115CookieButtonClick" />
+                    </StackPanel>
+                </RelativePanel>
+            </controls:SettingsCard>
+
+            <controls:SettingsCard Header="请求间隔"
+                                   Content="1s" />
+
+            <controls:SettingsCard Header="记录下载请求">
+                <StackPanel Orientation="Horizontal">
+                    <ToggleSwitch x:Name="IsRecordDownRequestToggleSwitch"
+                                  OnContent="开启"
+                                  OffContent="关闭"
+                                  IsOn="{x:Bind data:AppSettings.IsRecordDownRequest,Mode=TwoWay}" />
+
+                    <HyperlinkButton Content="清空"
+                                     Click="ClearDownRecordButton_Click"
+                                     RelativePanel.LeftOf="IsRecordDownRequestToggleSwitch"
+                                     Foreground="{ThemeResource DefaultTextForegroundThemeBrush}"
+                                     Opacity="0.5" />
+                </StackPanel>
+            </controls:SettingsCard>
+
+            <controls:SettingsCard Header="下载记录失效时间">
+                <StackPanel Orientation="Horizontal"
+                            Spacing="5">
+                    <TextBlock Opacity="0.3"
+                               Text="{x:Bind OverdueTimeNumberBox.Value,Converter={StaticResource NumberToLengthStrConverter} ,Mode=OneWay}"
+                               VerticalAlignment="Center" />
+                    <NumberBox x:Name="OverdueTimeNumberBox"
+                               SmallChange="60"
+                               LargeChange="3600"
+                               Minimum="0"
+                               Value="{x:Bind data:AppSettings.DownUrlOverdueTime,Mode=TwoWay}" />
+                </StackPanel>
+            </controls:SettingsCard>
+        </StackPanel>
+
+    </Grid>
+</Page>
