+++ Display/Views/Settings/CommonPage.xaml.cs
@@ -1,0 +1,17 @@
+using Microsoft.UI.Xaml.Controls;
+
+// To learn more about WinUI, the WinUI project structure,
+// and more about our project templates, see: http://aka.ms/winui-project-info.
+
+namespace Display.Views.Settings;
+
+/// <summary>
+/// An empty page that can be used on its own or navigated to within a Frame.
+/// </summary>
+public sealed partial class CommonPage : Page
+{
+    public CommonPage()
+    {
+        this.InitializeComponent();
+    }
+}
