+++ Display/Views/Settings/LoginPage.xaml
@@ -4,7 +4,6 @@
       xmlns:local="using:Display.Views.Settings"
     xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
     xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-    xmlns:customControl="using:Display.Control"
     mc:Ignorable="d"
     Background="{ThemeResource ApplicationPageBackgroundThemeBrush}">
 
@@ -48,7 +47,7 @@
         </Grid>
 
         <!--标题-->
-        <Grid
+        <Grid Grid.Row="0"
             x:Name="AppTitleBar"
             VerticalAlignment="Top"
             Height="28"
@@ -63,8 +62,7 @@
                         <ImageBrush ImageSource="/Assets/115/115favicon.ico"/>
                     </Border.Background>
                 </Border>
-                <TextBlock  x:Name="AppTitleTextBlock"
-                        Text="115"
+                <TextBlock Text="115"
                         TextWrapping="NoWrap"
                         FontSize="14"
                         Style="{StaticResource CaptionTextBlockStyle}" 
@@ -119,16 +117,16 @@
 
                     <Grid Height="200" Width="200">
                         <Image
-                                x:Name="QRCodeImage"
+                                x:Name="QrCodeImage"
                                 Height="200"
                                 Width="200"/>
 
-                        <Grid x:Name="imageGrid" Margin="15" CornerRadius="5" BorderThickness="1" BorderBrush="SkyBlue">
+                        <Grid x:Name="ImageGrid" Margin="15" CornerRadius="5" BorderThickness="1" BorderBrush="SkyBlue">
                             <TextBlock Text="等待刷新" HorizontalAlignment="Center" VerticalAlignment="Center" FontSize="20" CharacterSpacing="100"/>
                         </Grid>
 
 
-                        <Grid x:Name="RefreshQRCode_Grid" CornerRadius="5" Opacity="0" Margin="15" PointerEntered="RefreshQRCode_Grid_PointerEntered" PointerExited="RefreshQRCode_Grid_PointerExited" Tapped="RefreshQRCode_Grid_Tapped">
+                        <Grid x:Name="RefreshQrCodeGrid" CornerRadius="5" Opacity="0" Margin="15" PointerEntered="RefreshQRCode_Grid_PointerEntered" PointerExited="RefreshQRCode_Grid_PointerExited" Tapped="RefreshQRCode_Grid_Tapped">
                             <Grid.Background>
                                 <SolidColorBrush Color="AliceBlue" Opacity="0.9"/>
                             </Grid.Background>
@@ -137,14 +135,14 @@
                     </Grid>
 
                     <Grid Grid.Row="1">
-                        <TextBlock x:Name="sweepQRCodeTip" HorizontalAlignment="Center" VerticalAlignment="Center" Visibility="Collapsed">
+                        <TextBlock x:Name="SweepQrCodeTip" HorizontalAlignment="Center" VerticalAlignment="Center" Visibility="Collapsed">
                             <Run>使用</Run>
                             <Run Foreground="#e74c3c">115App</Run>
                             <Run>扫描二维码登录</Run>
                         </TextBlock>
 
                         <!--#62b84b-->
-                        <StackPanel x:Name="QRCodeConfirmTip" Visibility="Collapsed" Background="OrangeRed" Padding="6" CornerRadius="15" HorizontalAlignment="Center" VerticalAlignment="Center">
+                        <StackPanel x:Name="QrCodeConfirmTip" Visibility="Collapsed" Background="OrangeRed" Padding="6" CornerRadius="15" HorizontalAlignment="Center" VerticalAlignment="Center">
                             <TextBlock Foreground="White">
                             扫描成功 请在手机点确认以登录
                             </TextBlock>
@@ -163,18 +161,18 @@
                     Spacing="20">
                     <StackPanel>
                         <TextBox x:Name="CookieInputTextBox" Header="Cookie:" PlaceholderText="输入Cookie后提交" MinWidth="380" IsSpellCheckEnabled="False" TextChanged="CookieInputTextBox_TextChanged"/>
-                        <HyperlinkButton x:Name="checkCookieButton" Content="检查" HorizontalAlignment="Right" Click="CheckCookieButton_Click" IsEnabled="False">
+                        <HyperlinkButton x:Name="CheckCookieButton" Content="检查" HorizontalAlignment="Right" Click="CheckCookieButton_Click" IsEnabled="False">
                             <FlyoutBase.AttachedFlyout>
                                 <Flyout x:Name="CheckCookieFlyout"/>
                             </FlyoutBase.AttachedFlyout>
                         </HyperlinkButton>
 
-                        <TeachingTip x:Name="checkCookieTip"
-                                     Target="{x:Bind checkCookieButton}"
+                        <TeachingTip x:Name="CheckCookieTip"
+                                     Target="{x:Bind CheckCookieButton}"
                                      IsLightDismissEnabled="True"/>
 
                         <ContentDialog
-                            x:Name="termsOfUseContentDialog"
+                            x:Name="TermsOfUseContentDialog"
                             CloseButtonText="退出">
                         </ContentDialog>
 
@@ -186,7 +184,6 @@
             </Grid>
 
             <!--提示-->
-            <!--<Grid/>-->
         </Grid>
 
     </Grid>
