+++ Display/Views/Settings/PathPage.xaml
@@ -1,0 +1,49 @@
+<?xml version="1.0" encoding="utf-8"?>
+<Page
+    x:Class="Display.Views.Settings.PathPage"
+    xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+    xmlns:local="using:Display.Views.Settings"
+    xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+    xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+    xmlns:controls="using:CommunityToolkit.WinUI.Controls"
+    xmlns:options="using:Display.Models.Settings.Options"
+    mc:Ignorable="d"
+    Background="{ThemeResource ApplicationPageBackgroundThemeBrush}">
+
+    <Page.Resources>
+        <Style TargetType="controls:SettingsCard"
+               BasedOn="{StaticResource BaseSettingsCardStyle}" />
+
+        <StackLayout x:Key="SettingStackLayout"
+                     Spacing="{StaticResource SettingsCardPadding}"/>
+
+        <DataTemplate x:Key="SettingDataTemplate" x:DataType="options:SavePath">
+            <controls:SettingsCard Header="{x:Bind Name}"
+                                   ContentAlignment="Vertical"
+                                   HorizontalContentAlignment="Stretch">
+                <RelativePanel>
+                    <TextBox Width="600"
+                             Text="{x:Bind Path, Mode=TwoWay}" />
+                    <Button Content="修改"
+                            RelativePanel.LeftOf="OpenPathButton"
+                            Click="{x:Bind UpdatePathButtonClick}" />
+                    <Button x:Name="OpenPathButton"
+                            Click="{x:Bind OpenPathButtonClick}"
+                            Content="打开路径"
+                            RelativePanel.AlignRightWithPanel="True"
+                            Margin="3,0,0,0" />
+                </RelativePanel>
+            </controls:SettingsCard>
+        </DataTemplate>
+    </Page.Resources>
+
+
+    <StackPanel Style="{StaticResource SettingStackPanel}">
+        <ItemsRepeater Layout="{StaticResource SettingStackLayout}"
+                       ItemTemplate="{StaticResource SettingDataTemplate}"
+                       ItemsSource="{x:Bind _savePaths}">
+            
+        </ItemsRepeater>
+    </StackPanel>
+</Page>
