+++ Display/Views/Settings/SearchPage.xaml
@@ -1,0 +1,133 @@
+<?xml version="1.0" encoding="utf-8"?>
+<Page
+    x:Class="Display.Views.Settings.SearchPage"
+    xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+    xmlns:local="using:Display.Views.Settings"
+    xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+    xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+    xmlns:controls="using:CommunityToolkit.WinUI.Controls"
+    xmlns:data="using:Display.Models.Data"
+    mc:Ignorable="d"
+    Background="{ThemeResource ApplicationPageBackgroundThemeBrush}">
+
+    <Page.Resources>
+        <Style TargetType="Pivot"
+               BasedOn="{StaticResource SettingsPivotStyle}">
+            <Setter Property="HeaderTemplate">
+                <Setter.Value>
+                    <DataTemplate x:DataType="x:String">
+                        <TextBlock Text="{x:Bind}"
+                                   FontSize="14" />
+                    </DataTemplate>
+                </Setter.Value>
+            </Setter>
+        </Style>
+    </Page.Resources>
+
+    <StackPanel Style="{StaticResource SettingStackPanel}">
+
+        <controls:SettingsExpander Header="搜索资源">
+
+            <controls:SettingsExpander.Items>
+                <controls:SettingsCard ContentAlignment="Left">
+                    <Pivot>
+                        <PivotItem Header="搜索源">
+                            <StackPanel Spacing="5">
+                                <TextBlock Text="选择" />
+                                <StackPanel Orientation="Horizontal"
+                                            Spacing="10">
+                                    <ToggleButton Content="x1080x"
+                                                  IsChecked="{x:Bind data:AppSettings.IsUseX1080X,Mode=TwoWay}" />
+                                </StackPanel>
+
+                                <TextBlock Text="保存路径" />
+                                <RelativePanel>
+                                    <TextBox x:Name="SavePath115CidTextBlock"
+                                             PlaceholderText="从文件夹中选择。或者输入文件夹的cid，输入后点击保存"
+                                             Width="200"
+                                             Text="{x:Bind data:AppSettings.SavePath115Cid}" />
+                                    <TextBlock x:Name="SavePath115NameTextBlock"
+                                               Text="{x:Bind data:AppSettings.SavePath115Name}"
+                                               RelativePanel.RightOf="SavePath115CidTextBlock"
+                                               RelativePanel.AlignVerticalCenterWithPanel="True"
+                                               Margin="2,0,0,0" />
+
+                                    <StackPanel RelativePanel.AlignRightWithPanel="True"
+                                                Orientation="Horizontal"
+                                                Spacing="5">
+                                        <Button Content="从目录中选择"
+                                                Click="Selected115SavePathButtonClick" />
+                                        <Button Content="保存"
+                                                Click="Save115SavePathButtonClick" />
+                                    </StackPanel>
+                                </RelativePanel>
+                            </StackPanel>
+                        </PivotItem>
+
+                        <PivotItem Header="x1080x">
+                            <StackPanel Spacing="5">
+                                <RelativePanel>
+                                    <TextBlock x:Name="X1080XUrlTextBlock"
+                                               RelativePanel.AlignVerticalCenterWithPanel="True"
+                                               Width="70">地址</TextBlock>
+                                    <TextBox x:Name="X1080UrlTextBox"
+                                             Width="600"
+                                             RelativePanel.RightOf="X1080XUrlTextBlock"
+                                             Text="{x:Bind data:AppSettings.X1080XBaseUrl}" />
+                                    <Button Click="X1080XBaseUrlChange"
+                                            Content="修改"
+                                            RelativePanel.AlignRightWithPanel="True" />
+                                </RelativePanel>
+
+                                <RelativePanel>
+                                    <TextBlock x:Name="X1080XCookieTextBlock"
+                                               RelativePanel.AlignVerticalCenterWithPanel="True"
+                                               Width="70"
+                                               Text="Cookie" />
+                                    <TextBox Width="600"
+                                             x:Name="X1080XCookieTextBox"
+                                             RelativePanel.RightOf="X1080XCookieTextBlock"
+                                             IsSpellCheckEnabled="False"
+                                             Text="{x:Bind data:AppSettings.X1080XCookie}">
+                                        <ToolTipService.ToolTip>
+                                            <TextBlock>
+                                                        搜索资源需要登陆后的Cookie
+                                            </TextBlock>
+                                        </ToolTipService.ToolTip>
+                                    </TextBox>
+                                    <Button Click="X1080XCookieChange"
+                                            Content="修改"
+                                            RelativePanel.AlignRightWithPanel="True" />
+                                </RelativePanel>
+
+                                <RelativePanel>
+                                    <TextBlock x:Name="X1080XuaTextBlock"
+                                               RelativePanel.AlignVerticalCenterWithPanel="True"
+                                               Width="70">UserAgent</TextBlock>
+                                    <TextBox x:Name="X1080XuaTextBox"
+                                             Width="600"
+                                             RelativePanel.RightOf="X1080XuaTextBlock"
+                                             Text="{x:Bind data:AppSettings.X1080XUa}"
+                                             IsSpellCheckEnabled="False">
+                                        <ToolTipService.ToolTip>
+                                            <TextBlock>
+                                                        需要与登录端的UserAgent匹配
+                                            </TextBlock>
+                                        </ToolTipService.ToolTip>
+                                    </TextBox>
+                                    <Button Click="X1080XUserAgentChange"
+                                            Content="修改"
+                                            RelativePanel.AlignRightWithPanel="True" />
+                                </RelativePanel>
+
+                            </StackPanel>
+                        </PivotItem>
+                    </Pivot>
+                </controls:SettingsCard>
+
+            </controls:SettingsExpander.Items>
+        </controls:SettingsExpander>
+
+    </StackPanel>
+</Page>
