+++ Display/Views/Settings/SpiderPage.xaml
@@ -1,0 +1,73 @@
+<?xml version="1.0" encoding="utf-8"?>
+<Page
+    x:Class="Display.Views.Settings.SpiderPage"
+    xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+    xmlns:local="using:Display.Views.Settings"
+    xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+    xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+    xmlns:controls="using:CommunityToolkit.WinUI.Controls"
+    xmlns:options="using:Display.Models.Settings.Options"
+    mc:Ignorable="d"
+    Background="{ThemeResource ApplicationPageBackgroundThemeBrush}">
+
+
+    <StackPanel Style="{StaticResource SettingStackPanel}">
+        <controls:SettingsExpander Header="搜刮信息"
+                                   ItemsSource="{x:Bind _spiders}">
+
+            <controls:SettingsExpander.ItemTemplate>
+                <DataTemplate x:DataType="options:Spider">
+                    <controls:SettingsCard ContentAlignment="Vertical"
+                                           HorizontalContentAlignment="Stretch">
+                        <StackPanel Spacing="8">
+                            <RelativePanel BorderBrush="{ThemeResource ControlElevationBorderBrush}"
+                                           BorderThickness="0,0,0,1">
+                                <TextBlock Text="{x:Bind Instance.Name}"
+                                           Style="{StaticResource BodyStrongTextBlockStyle}"
+                                           RelativePanel.AlignVerticalCenterWithPanel="True" />
+                                <ToggleSwitch OnContent="启动"
+                                              OffContent="关闭"
+                                              IsOn="{x:Bind Instance.IsOn,Mode=TwoWay}"
+                                              Unloaded="SpiderToggleButton_Unchecked"
+                                              RelativePanel.AlignRightWithPanel="True" />
+                            </RelativePanel>
+
+
+                            <RelativePanel>
+                                <TextBox Width="600"
+                                         Header="地址"
+                                         Text="{x:Bind Instance.BaseUrl, Mode=OneWay}" />
+                                <Button Content="检查"
+                                        IsEnabled="False"
+                                        RelativePanel.AlignRightWithPanel="True"
+                                        RelativePanel.AlignBottomWithPanel="True" />
+                            </RelativePanel>
+
+                            <RelativePanel Visibility="{x:Bind Instance.IsCookieEnable}">
+                                <TextBox Width="600"
+                                         Header="Cookie"
+                                         Text="{x:Bind Cookie, Mode=TwoWay}" />
+                                <Button Content="保存"
+                                        Click="{x:Bind UpdateCookie}"
+                                        RelativePanel.AlignRightWithPanel="True"
+                                        RelativePanel.AlignBottomWithPanel="True" />
+                            </RelativePanel>
+
+
+                            <RelativePanel>
+                                <TextBlock Text="请求间隔" />
+                                <TextBlock RelativePanel.AlignRightWithPanel="True">
+                                                <Run Text="{x:Bind Instance.DelayRanges.Item1}" />
+                                                ~
+                                                <Run Text="{x:Bind Instance.DelayRanges.Item2}" />
+                                                s
+                                </TextBlock>
+                            </RelativePanel>
+                        </StackPanel>
+                    </controls:SettingsCard>
+                </DataTemplate>
+            </controls:SettingsExpander.ItemTemplate>
+        </controls:SettingsExpander>
+    </StackPanel>
+</Page>
